Refactor controller
